Inline format_period_slug, move format_date_range to date_utils
