+++ git_dev_metrics/cli/wizards/summary_wizard.py
@@ -5,7 +5,7 @@ import typer
 
 from ...metrics.loader import load_snapshot_for_months
 from ...metrics.printer import ConsolePrinter
-from ...utils._snapshot_format import format_date_range, format_period_slug
+from ...utils.date_utils import format_date_range
 from ._wizard import _prompt_months, pick_months
 
 YearMonth = tuple[int, int]
@@ -23,7 +23,7 @@ def summary_wizard(
         typer.secho("No PRs in selected months.", fg=typer.colors.RED, err=True)
         raise typer.Exit(code=1)
     first, last = selected[0], selected[-1]
-    period_slug = format_period_slug(
-        f"{first[0]:04d}-{first[1]:02d}", f"{last[0]:04d}-{last[1]:02d}"
+    period_slug = f"{first[0]:04d}-{first[1]:02d}-to-{last[0]:04d}-{last[1]:02d}"
+    ConsolePrinter().print_combined_metrics(
+        snapshot, period_slug, format_date_range(snapshot.period)
     )
-    ConsolePrinter().print_combined_metrics(snapshot, period_slug, format_date_range(snapshot))
